pyproject.toml
[tool.poetry]
name = "mentoria backend"
version = "0.1.0"
description = "Mentoria"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.0"
uvicorn = "^0.28.0"
python-decouple = "^3.8"
pyodbc = "^5.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
